Rename
This contains a glimpse of the whole project. It shows the object of the first research step. This notebook uses the median as a tool to inspect the double exponential distribution. This is a not good method to visualize the dataset. Further work uses a density plot to show the distribution This notebook serves as a baby step.